Added sourcemap support
In this commit, we add support for the sourceMap compile option. This
option makes coffee.compile return an object with both the generated JS
and a JSON V3 source map.

Because we don't want to host the map seperately, we base64-encode both the map
*and the original coffeescript source* into a data URL. This strategy avoids a
separate render pass and works around bevry/docpad#765